Rename to singular to match other class names
