+++ rules-tests/DeadCode/Rector/If_/RemoveDeadIfBlockRector/Fixture/if_else_no_stmt.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlocksRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlockRector\Fixture;
 
 class IfElseNoStmt
 {
@@ -17,7 +17,7 @@ class IfElseNoStmt
 -----
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlocksRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlockRector\Fixture;
 
 class IfElseNoStmt
 {
