+++ rules-tests/DeadCode/Rector/If_/RemoveDeadIfBlockRector/Fixture/if_no_stmt_with_else.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlocksRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlockRector\Fixture;
 
 class IfNoStmtWithElse
 {
@@ -17,7 +17,7 @@ class IfNoStmtWithElse
 -----
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlocksRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlockRector\Fixture;
 
 class IfNoStmtWithElse
 {
