+++ rules-tests/DeadCode/Rector/If_/RemoveDeadIfBlockRector/Fixture/skip_property_fetch.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlocksRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlockRector\Fixture;
 
 class SkipPropertyFetch
 {
