+++ rules-tests/DeadCode/Rector/If_/RemoveDeadIfBlockRector/Fixture/skip_used_in_next_stmt2.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlocksRector\Fixture;
+namespace Rector\Tests\DeadCode\Rector\If_\RemoveDeadIfBlockRector\Fixture;
 
 class SkipUsedInNextStmt2
 {
