--- rules-tests/DeadCode/Rector/If_/RemoveDeadIfBlocksRector/config/configured_rule.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\DeadCode\Rector\If_\RemoveDeadIfBlocksRector;
-
-return RectorConfig::configure()
-    ->withRules([RemoveDeadIfBlocksRector::class]);
