+++ src/Config/Level/DeadCodeLevel.php
@@ -41,7 +41,7 @@ use Rector\DeadCode\Rector\FunctionLike\NarrowWideUnionReturnTypeRector;
 use Rector\DeadCode\Rector\FunctionLike\RemoveDeadReturnRector;
 use Rector\DeadCode\Rector\If_\ReduceAlwaysFalseIfOrRector;
 use Rector\DeadCode\Rector\If_\RemoveAlwaysTrueIfConditionRector;
-use Rector\DeadCode\Rector\If_\RemoveDeadIfBlocksRector;
+use Rector\DeadCode\Rector\If_\RemoveDeadIfBlockRector;
 use Rector\DeadCode\Rector\If_\RemoveDeadInstanceOfRector;
 use Rector\DeadCode\Rector\If_\RemoveTypedPropertyDeadInstanceOfRector;
 use Rector\DeadCode\Rector\If_\RemoveUnusedNonEmptyArrayBeforeForeachRector;
@@ -130,7 +130,7 @@ final class DeadCodeLevel
 
         RemoveDeadCatchRector::class,
         RemoveDeadTryCatchRector::class,
-        RemoveDeadIfBlocksRector::class,
+        RemoveDeadIfBlockRector::class,
         RemoveDeadIfForeachForRector::class,
         RemoveConditionExactReturnRector::class,
         RemoveDeadStmtRector::class,
